Add proper identity propagation to WebSockets
Fixes #16602
