+++ bom/application/pom.xml
@@ -30,7 +30,7 @@
         <opentelemetry.version>1.6.0</opentelemetry.version>
         <opentelemetry-alpha.version>1.6.0-alpha</opentelemetry-alpha.version>
         <jaeger.version>1.6.0</jaeger.version>
-        <quarkus-http.version>4.1.1</quarkus-http.version>
+        <quarkus-http.version>4.1.2</quarkus-http.version>
         <micrometer.version>1.7.4</micrometer.version>
         <google-auth.version>0.22.0</google-auth.version>
         <microprofile-config-api.version>2.0</microprofile-config-api.version>
