--- extensions/websockets/server/deployment/src/test/java/io/quarkus/undertow/websockets/test/EchoService.java
@@ -1,11 +1,0 @@
-package io.quarkus.undertow.websockets.test;
-
-import javax.enterprise.context.RequestScoped;
-
-@RequestScoped
-public class EchoService {
-
-    public String echo(String msg) {
-        return msg;
-    }
-}
