--- extensions/websockets/server/deployment/src/test/java/io/quarkus/undertow/websockets/test/WebsocketDevModeTestCase.java
@@ -1,87 +1,0 @@
-package io.quarkus.undertow.websockets.test;
-
-import java.net.URI;
-import java.util.concurrent.LinkedBlockingDeque;
-import java.util.concurrent.TimeUnit;
-import java.util.function.Supplier;
-
-import javax.websocket.ClientEndpointConfig;
-import javax.websocket.ContainerProvider;
-import javax.websocket.Endpoint;
-import javax.websocket.EndpointConfig;
-import javax.websocket.MessageHandler;
-import javax.websocket.Session;
-
-import org.jboss.shrinkwrap.api.ShrinkWrap;
-import org.jboss.shrinkwrap.api.spec.JavaArchive;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.RegisterExtension;
-
-import io.quarkus.test.QuarkusDevModeTest;
-import io.quarkus.test.common.http.TestHTTPResource;
-
-/**
- * smoke test that websockets work as expected in dev mode
- */
-public class WebsocketDevModeTestCase {
-
-    @TestHTTPResource("echo")
-    URI echoUri;
-
-    @RegisterExtension
-    public static final QuarkusDevModeTest test = new QuarkusDevModeTest()
-            .setArchiveProducer(new Supplier<JavaArchive>() {
-                @Override
-                public JavaArchive get() {
-                    return ShrinkWrap.create(JavaArchive.class)
-                            .addClasses(EchoWebSocket.class, EchoService.class);
-                }
-            });
-
-    @Test
-    public void testWebsocketHotReplacement() throws Exception {
-
-        LinkedBlockingDeque<String> message = new LinkedBlockingDeque<>();
-        Session session = ContainerProvider.getWebSocketContainer().connectToServer(new Endpoint() {
-            @Override
-            public void onOpen(Session session, EndpointConfig endpointConfig) {
-                session.addMessageHandler(new MessageHandler.Whole<String>() {
-                    @Override
-                    public void onMessage(String s) {
-                        message.add(s);
-                    }
-                });
-                session.getAsyncRemote().sendText("hello");
-            }
-        }, ClientEndpointConfig.Builder.create().build(), echoUri);
-
-        try {
-            Assertions.assertEquals("hello", message.poll(20, TimeUnit.SECONDS));
-        } finally {
-            session.close();
-        }
-
-        test.modifySourceFile(EchoService.class, (s) -> s.replace("return msg;", "return \"changed:\" + msg;"));
-
-        session = ContainerProvider.getWebSocketContainer().connectToServer(new Endpoint() {
-            @Override
-            public void onOpen(Session session, EndpointConfig endpointConfig) {
-                session.addMessageHandler(new MessageHandler.Whole<String>() {
-                    @Override
-                    public void onMessage(String s) {
-                        message.add(s);
-                    }
-                });
-                session.getAsyncRemote().sendText("hello");
-            }
-        }, ClientEndpointConfig.Builder.create().build(), echoUri);
-
-        try {
-            Assertions.assertEquals("changed:hello", message.poll(20, TimeUnit.SECONDS));
-        } finally {
-            session.close();
-        }
-
-    }
-}
