+++ extensions/websockets/server/deployment/src/test/java/io/quarkus/websockets/test/EchoService.java
@@ -1,0 +1,11 @@
+package io.quarkus.websockets.test;
+
+import javax.enterprise.context.RequestScoped;
+
+@RequestScoped
+public class EchoService {
+
+    public String echo(String msg) {
+        return msg;
+    }
+}
